Move constants into a dedicated submodule
